requirements.txt
torch_scatter==2.1.2
numpy==1.26.4
scikit-learn==1.0.2
torch_sparse==0.4.3
torch_cluster==1.4.5
torch_geometric==2.5.3
tqdm==4.40.0
joblib==0.14.1

